2017/11/14 19:43 zsh done

+++ app/src/main/java/zzbcar/cckj/com/nzzb/bean/CarDefaultBean.java
@@ -102,7 +102,7 @@ public class CarDefaultBean {
         private String pics;
         private String plateNo;
         private double price;
-        private int receivePercent;
+        private double receivePercent;
         private int seatNum;
         private int series;
         private String seriesName;
@@ -303,11 +303,11 @@ public class CarDefaultBean {
             this.price = price;
         }
 
-        public int getReceivePercent() {
+        public double getReceivePercent() {
             return receivePercent;
         }
 
-        public void setReceivePercent(int receivePercent) {
+        public void setReceivePercent(double receivePercent) {
             this.receivePercent = receivePercent;
         }
